Update packages to avoid conflict
