+++ src/ca/team2706/frc/autonomous/AutoBuilder.java
@@ -1,0 +1,20 @@
+package ca.team2706.frc.autonomous;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class AutoBuilder {
+	private List<AutoCommand> commands = new ArrayList<AutoCommand>();
+	
+	public AutoBuilder() {
+		
+	}
+	
+	public void add(AutoCommand autoCommand) {
+		commands.add(autoCommand);
+	}
+	
+	public AutoCommand[] toArray() {
+		return commands.toArray(new AutoCommand[commands.size()]);
+	}
+}
