+++ src/ca/team2706/frc/autonomous/AutoCommand.java
@@ -1,0 +1,7 @@
+package ca.team2706.frc.autonomous;
+
+public interface AutoCommand {
+	public void initialize();
+	public boolean tick();
+	public void cleanup();
+}
