+++ src/ca/team2706/frc/autonomous/modes/BasicRobotSetMode.java
@@ -1,0 +1,21 @@
+package ca.team2706.frc.autonomous.modes;
+
+import ca.team2706.frc.autonomous.AutoBuilder;
+import ca.team2706.frc.autonomous.AutoCommand;
+import ca.team2706.frc.autonomous.AutoMode;
+import ca.team2706.frc.autonomous.commands.GearShift;
+
+public class BasicRobotSetMode extends AutoMode {
+	//TODO: Test this on practice field. These are bogus values.
+
+	@Override
+	protected AutoCommand[] initializeCommands() {
+		// This assumes that we start in the staging zone parallel to the closest scoring platform
+		AutoBuilder builder = new AutoBuilder();
+
+		// Shift to low gear
+		builder.add(new GearShift(false));
+		
+		return builder.toArray();
+	}
+}
