+++ src/ca/team2706/frc/autonomous/modes/DriveForwardMode.java
@@ -1,0 +1,28 @@
+package ca.team2706.frc.autonomous.modes;
+
+import ca.team2706.frc.autonomous.AutoBuilder;
+import ca.team2706.frc.autonomous.AutoCommand;
+import ca.team2706.frc.autonomous.AutoMode;
+import ca.team2706.frc.autonomous.commands.DriveStraight;
+import ca.team2706.frc.autonomous.commands.GearShift;
+
+public class DriveForwardMode extends AutoMode {
+	
+	//TODO: Test this on practice field. These are bogus values.
+	private static final double STRAIGHT_DISTANCE = 9.45; // TODO Change this number!!!!!
+//	private static final int SIMPLE_FORKLIFT_COUNTER = 50;
+	
+	@Override
+	protected AutoCommand[] initializeCommands() {
+		// Alignment: Back of Robot with Staging Zone 
+		AutoBuilder builder = new AutoBuilder();
+		
+		// Shift to low gear
+		builder.add(new GearShift(false));
+		
+		// Drive forward into the zone.
+		builder.add(new DriveStraight(STRAIGHT_DISTANCE));
+		
+		return builder.toArray();
+	}
+}
