--- src/com/owcrobots/frc/autonomous/AutoBuilder.java
@@ -1,20 +1,0 @@
-package com.owcrobots.frc.autonomous;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class AutoBuilder {
-	private List<AutoCommand> commands = new ArrayList<AutoCommand>();
-	
-	public AutoBuilder() {
-		
-	}
-	
-	public void add(AutoCommand autoCommand) {
-		commands.add(autoCommand);
-	}
-	
-	public AutoCommand[] toArray() {
-		return commands.toArray(new AutoCommand[commands.size()]);
-	}
-}
