--- src/com/owcrobots/frc/autonomous/AutoCommand.java
@@ -1,7 +1,0 @@
-package com.owcrobots.frc.autonomous;
-
-public interface AutoCommand {
-	public void initialize();
-	public boolean tick();
-	public void cleanup();
-}
