--- src/com/owcrobots/frc/controls/EEncoder.java
@@ -1,22 +1,0 @@
-package com.owcrobots.frc.controls;
-
-import edu.wpi.first.wpilibj.Encoder;
-
-public class EEncoder extends Encoder {
-	
-	private int offset = 0;
-
-	public EEncoder(int aChannel, int bChannel) {
-		super(aChannel, bChannel);
-	}
-	
-	public void reset(int offset) {
-		this.offset = offset;
-		super.reset();
-	}
-	
-	public int get() {
-		return super.get() + offset;
-	}
-
-}
